+++ src/lib.rs
@@ -1,3 +1,4 @@
 mod filesystem_utils;
+mod handlebars_helper;
 mod parabuilder;
 pub use parabuilder::{CompliationErrorHandlingMethod, Parabuilder, RunMethod};